require units, cleanup init (#1222)
